src/requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
praw==7.7.1
prawcore==2.4.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
SQLAlchemy==2.0.21
typing_extensions==4.8.0
update-checker==0.18.0
urllib3==2.0.6
websocket-client==1.6.4
Werkzeug==3.0.0
